+++ src/LuminaSupplemental.Excel/Model/SubmarineDrop.cs
@@ -17,15 +17,24 @@ namespace LuminaSupplemental.Excel.Model
     {
         [Name("ItemId")] public uint ItemId { get; set; }
         [Name("SubmarineExplorationId")] public uint SubmarineExplorationId { get; set; }
+        [Name("LootTier")] public byte LootTier { get; set; }
+        [Name("Min")] public uint Min { get; set; }
+        [Name("Max")] public uint Max { get; set; }
+        [Name("Probability")] public decimal Probability { get; set; }
+
 
         public RowRef< SubmarineExploration > SubmarineExploration;
 
         public RowRef< Item > Item;
 
-        public SubmarineDrop(uint itemId, uint submarineExplorationId )
+        public SubmarineDrop(uint itemId, uint submarineExplorationId, byte lootTier, uint min, uint max, decimal probability )
         {
             ItemId = itemId;
             SubmarineExplorationId = submarineExplorationId;
+            LootTier = lootTier;
+            Min = min;
+            Max = max;
+            Probability = probability;
         }
 
         public SubmarineDrop()
@@ -37,6 +46,10 @@ namespace LuminaSupplemental.Excel.Model
         {
             ItemId = uint.Parse( lineData[ 0 ] );
             SubmarineExplorationId = uint.Parse( lineData[ 1 ] );
+            LootTier = byte.Parse( lineData[ 2 ], CultureInfo.InvariantCulture );
+            Min = uint.Parse( lineData[ 3 ], CultureInfo.InvariantCulture );
+            Max = uint.Parse( lineData[ 4 ], CultureInfo.InvariantCulture );
+            Probability = decimal.Parse( lineData[ 5 ], CultureInfo.InvariantCulture );
         }
 
         public string[] ToCsv()