+++ CHANGELOG.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed+
 
+## [2.3.6] - 2025-06-01
+
+### Changed
+
+- Dungeon sheets updated for 7.25
+- Submarine drops now come from submersible-loot-db
+
 ## [2.3.5] - 2025-06-01
 
 ### Fixed